processor/requirements.txt
openai-whisper
ffmpeg-python
numpy
torch 